fix(cursor): add poll timestamp and extract shared mapper logic
- Add Timestamp to poll-synthesized AgentStatusWebhook for parity
  with webhook output path
- Extract baseEventSections into base.ts following Claude integration
  pattern, removing duplication from launchAgent.ts and getDailyUsage.ts
